refactor: move to curve-common/features subdir
