+++ apps/loan/src/pages/_app.tsx
@@ -5,7 +5,7 @@ import delay from 'lodash/delay'
 import 'intersection-observer'
 import 'focus-visible'
 import '@/globals.css'
-import { connectWalletLocales } from '@/common/shared/features/connect-wallet/lib'
+import { connectWalletLocales } from '@/common/shared/features/features/connect-wallet/lib'
 import { useCallback, useEffect, useState } from 'react'
 import { HashRouter } from 'react-router-dom'
 import { REFRESH_INTERVAL } from '@/constants'
@@ -22,7 +22,7 @@ import { isMobile, removeExtraSpaces } from '@/utils/helpers'
 import { getStorageValue } from '@/utils/storage'
 import { getLocaleFromUrl } from '@/utils/utilsRouter'
 import type { AppProps } from 'next/app'
-import { initOnboard } from '@/common/shared/features/connect-wallet/lib/init'
+import { initOnboard } from '@/common/shared/features/features/connect-wallet/lib/init'
 
 i18n.load({ en: messagesEn })
 i18n.activate('en')
