+++ packages/curve-common/src/features/connect-wallet/lib/init.ts
