+++ packages/curve-common/src/features/connect-wallet/lib/locales/index.ts
