+++ packages/curve-common/src/features/connect-wallet/lib/locales/zh-Hant.ts
